Move creating new rideRequest from rides to ridesRequest new, having more separated cruds
