+++ src/app/Rides/Update.elm
@@ -1,11 +1,10 @@
 module Rides.Update exposing (init, update)
 
-import Common.IdentifiedList exposing (findById, mapIfId)
 import Common.Response exposing (Response(..))
 import Model as Root exposing (Msg(..))
 import Return exposing (Return, return)
 import Rides.Model exposing (..)
-import Rides.Ports exposing (encodeRide, rideRequest, ridesList)
+import Rides.Ports exposing (ridesList)
 import UrlRouter.Model exposing (Msg(..))
 import UrlRouter.Routes exposing (Page(..), pathParser)
 
@@ -41,28 +40,3 @@ updateRides msg collection =
     case msg of
         UpdateRides response ->
             return { list = response } Cmd.none
-
-        Submit rideId groupId ->
-            updateRide rideId
-                (\ride ->
-                    return { ride | rideRequest = Loading } <|
-                        rideRequest (encodeRide groupId ride)
-                )
-                collection
-
-        RideRequestResponse rideId response ->
-            updateRide rideId
-                (\ride -> return { ride | rideRequest = response } Cmd.none)
-                collection
-
-
-updateRide : String -> (Model -> ( Model, Cmd Rides.Model.Msg )) -> Collection -> Return Rides.Model.Msg Collection
-updateRide id f collection =
-    case collection.list of
-        Success list ->
-            mapIfId id f (\model -> return model Cmd.none) list
-                |> Return.sequence
-                |> Return.map (\list -> { list = Success list })
-
-        _ ->
-            return collection Cmd.none
