+++ src/app/RidesRequests/Update.elm
@@ -1,29 +1,33 @@
 module RidesRequests.Update exposing (init, update)
 
-import Common.Response exposing (Response(..))
+import Common.Response as Response exposing (Response(..))
 import Model as Root exposing (Msg(..))
 import Return exposing (Return, return)
-import RidesRequests.Model as RidesRequests exposing (Model, Msg(..))
-import RidesRequests.Ports exposing (fetchRideRequest)
+import RidesRequests.Model as RidesRequests exposing (Collection, Msg(..))
+import RidesRequests.Ports exposing (createRideRequest, encodeRideRequest, fetchRideRequest)
 import UrlRouter.Model exposing (Msg(..))
 import UrlRouter.Routes exposing (Page(..), pathParser)
 
 
-init : Model
+init : Collection
 init =
-    Empty
+    { list = Empty
+    , new =
+        { response = Empty
+        }
+    }
 
 
-update : Root.Msg -> Model -> Return RidesRequests.Msg Model
-update msg model =
+update : Root.Msg -> Collection -> Return RidesRequests.Msg Collection
+update msg collection =
     case msg of
         MsgForRidesRequests msg_ ->
-            updateRidesRequests msg_ model
+            updateRidesRequests msg_ collection
 
         MsgForUrlRouter (UrlChange location) ->
             case pathParser location of
                 Just (RideRequestPage groupId rideId fromUserId rideRequestId) ->
-                    return Loading
+                    return { collection | list = Loading }
                         (fetchRideRequest
                             { groupId = groupId
                             , rideId = rideId
@@ -33,14 +37,21 @@ update msg model =
                         )
 
                 _ ->
-                    return model Cmd.none
+                    return collection Cmd.none
 
         _ ->
-            return model Cmd.none
+            return collection Cmd.none
 
 
-updateRidesRequests : RidesRequests.Msg -> Model -> Return RidesRequests.Msg Model
-updateRidesRequests msg model =
+updateRidesRequests : RidesRequests.Msg -> Collection -> Return RidesRequests.Msg Collection
+updateRidesRequests msg collection =
     case msg of
         FetchedRideRequest response ->
-            return response Cmd.none
+            return { collection | list = Response.map (\ride -> [ ride ]) response } Cmd.none
+
+        CreateRideRequest ride ->
+            return { collection | new = { response = Loading } }
+                (createRideRequest <| encodeRideRequest ride)
+
+        CreateRideRequestResponse rideId response ->
+            return { collection | new = { response = response } } Cmd.none
