+++ src/app/Stylesheets.elm
@@ -6,7 +6,6 @@ import Layout.Styles
 import Login.Styles
 import Notifications.Styles
 import Profile.Styles
-import Rides.Ride.Styles
 import Rides.Styles
 
 
@@ -22,7 +21,6 @@ fileStructure =
                 , Login.Styles.styles
                 , Rides.Styles.styles
                 , Notifications.Styles.styles
-                , Rides.Ride.Styles.styles
                 , Profile.Styles.styles
                 , Groups.Styles.styles
                 ]
