+++ src/js/firebase/ridesRequests.js
@@ -2,6 +2,8 @@ var success = require("./helpers").success;
 var error = require("./helpers").error;
 
 module.exports = function(firebase, app) {
+  var getProfile = require("./profile").getProfile(firebase, app);
+
   app.ports.fetchRideRequest.subscribe(function(ids) {
     firebase
       .database()
@@ -36,6 +38,37 @@ module.exports = function(firebase, app) {
       })
       .catch(function(err) {
         app.ports.fetchRideRequestResponse.send(error(err.message));
+      });
+  });
+
+  app.ports.createRideRequest.subscribe(function(rideRequest) {
+    getProfile()
+      .then(function(profile) {
+        return firebase
+          .database()
+          .ref(
+            "ridesRequests/" +
+              rideRequest.groupId +
+              "/" +
+              rideRequest.rideId +
+              "/" +
+              rideRequest.toUserId +
+              "/" +
+              firebase.auth().currentUser.uid
+          )
+          .push({ profile: profile.val() });
+      })
+      .then(function() {
+        app.ports.createRideRequestResponse.send({
+          rideId: rideRequest.rideId,
+          response: success(true)
+        });
+      })
+      .catch(function(err) {
+        app.ports.createRideRequestResponse.send({
+          rideId: rideRequest.rideId,
+          response: error(err.message)
+        });
       });
   });
 };
